More changes for NGO onboarding and generic service creations.Fixed many issues code

+++ hni-user/src/main/java/org/hni/user/service/UserService.java
@@ -39,7 +39,7 @@ public interface UserService extends BaseService<User> {
 
 	User byEmailAddress(String emailAddress);
 
-	default User register(User user, Integer userType) {
+	default User register(User user, Long userType) {
 		return user;
 	}
